src/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiowebsocket==1.0.0.dev2
anyio==3.5.0
asgiref==3.5.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
bcrypt==4.0.1
blueprint==3.4.2
cachelib==0.10.2
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
comm==0.1.2
cssselect==1.1.0
cycler==0.11.0
cryptography
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.4
djlint==0.7.4
enum34==1.1.6
executing==1.2.0
ffmpeg-python
fastapi==0.74.1
filelock==3.4.2
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Dance==6.2.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-MySQLdb==1.0.1
Flask-OAuthlib==0.9.6
flask-paginate==2024.4.12
Flask-SQLAlchemy==2.5.1
fonttools==4.31.1
frozenlist==1.3.0
gevent==21.12.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
google-cloud-storage
greenlet==1.1.2
gspread==5.7.2
gunicorn==20.1.0
h11==0.13.0
httpcore==1.0.5
httplib2==0.21.0
httpx==0.27.0
idna==3.7
importlib-metadata==4.11.0
ipykernel==6.21.3
ipython==8.11.0
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.0.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.0
lxml==4.8.0
MarkupSafe==2.1.2
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
matplotlib==3.5.1
matplotlib-inline==0.1.6
mercadopago==2.2.1
multidict==6.0.2
mysqlclient==2.1.0
nest-asyncio==1.5.6
numpy==1.22.3
oauth2client==4.1.3
oauthlib==2.1.0
openai==1.72.0
packaging==21.3
pandas==2.0.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
paypal-checkout-serversdk==1.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyDrive==1.3.1
Pygments==2.14.0
pyhomebroker==0.48.1
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.7
pyquery==1.4.3
pyRofex==0.5.0rc1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==25.0.2
redis==5.0.1
regex==2022.1.18
requests==2.31.0
requests-oauthlib==1.1.0
rsa==4.9
redis==5.0.1
schedule==1.2.1
signalr-client-threads==0.0.12
simplejson==3.10.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
sseclient==0.0.27
stack-data==0.6.2
starlette==0.17.1
tomlkit==0.8.0
tornado==6.2
tqdm==4.62.3
traitlets==5.9.0
typing-extensions>=4.11,<5
tzdata==2024.1
upgrade-requirements==1.7.0
uritemplate==4.1.1
urllib3==1.24.3
URLObject==2.4.3
uvicorn==0.17.5
virtualenv==20.13.1
wcwidth==0.2.6
websocket==0.2.1
websocket-client==0.57.0
websockets==10.4
Werkzeug==2.2.3
xlwings==0.27.1
yarl==1.7.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0


